requirements.txt
botbuilder-core==4.14.5
botbuilder-schema==4.14.5
aiohttp==3.9.5
gunicorn==21.2.0
